T_P_OLD/requirements.txt
Flask
requests
python-dotenv
google-generativeai  